requirements/base.txt
uvicorn==0.30.1
fastapi==0.111.1
pydantic-settings==2.3.4
asyncpg==0.29.0
alembic==1.13.2
python-json-logger==2.0.7
pytest==8.3.1
greenlet==3.0.3
email_validator==2.2.0
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
redis==5.0.7
async-asgi-testclient=1.4.11
pytest-asyncio==0.23.8